UPDATED: Fixed all the makefiles to remove the debug stuff that should never have been committed but was by accident UPDATED: Fixed the fixedPoint system's includes up to get rid of the non-needed and non-available-on-Windows strings. UPDATED: Fixed a bug with the module mixer Mix() function which was causing Spectrometer problems rightly due to the size/signage error that was happening UPDATED: Fixed up the visual studio project so things all build again UPDATED: Fully switched over to the LinearSlideDown() function, removing the LinearSlideDownTable

+++ libAudio/genericModule/genericModule.h
@@ -318,7 +318,7 @@ public:
 	const char *GetTitle();
 	uint8_t GetChannels();
 	void InitMixer(FileInfo *p_FI);
-	long Mix(uint8_t *Buffer, uint32_t BuffLen);
+	int32_t Mix(uint8_t *Buffer, uint32_t BuffLen);
 };
 
 #endif /*__GenericModule_H__*/
